requirements.txt
# Web Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Database
sqlalchemy==2.0.36
alembic==1.14.0

# Testing
hypothesis==6.122.3
pytest==8.3.4
pytest-asyncio==0.24.0

# Data Processing
pandas==2.2.3
numpy==2.1.3

# Dashboard
streamlit==1.40.2
plotly==5.24.1
# folium==0.18.0
# streamlit-folium==0.23.1

# AI/ML
openai==1.57.2

# Utilities
python-dotenv==1.0.1
cryptography==44.0.0
pydantic==2.10.3
httpx==0.28.1
filelock==3.16.1
